move logging to config file
