+++ oidc_client/config/__init__.py
@@ -1,11 +1,16 @@
 """OIDC Client Configuration."""
 
 import os
+import logging
 
 from pathlib import Path
 from configparser import ConfigParser
 from collections import namedtuple
 from distutils.util import strtobool
+
+formatting = '[%(asctime)s][%(name)s][%(process)d %(processName)s][%(levelname)-8s] (L:%(lineno)s) %(module)s | %(funcName)s: %(message)s'
+logging.basicConfig(level=logging.DEBUG if bool(strtobool(os.environ.get('DEBUG', 'False'))) else logging.INFO, format=formatting)
+LOG = logging.getLogger("oidc")
 
 
 def parse_config_file(path):
