+++ oidc_client/endpoints/token.py
@@ -1,7 +1,7 @@
 """Token Endpoint."""
 
 from ..utils.utils import get_from_session
-from ..utils.logging import LOG
+from ..config import LOG
 
 
 async def token_request(request):
